feat(waylandIdle): move functionality into a Vencord plugin
Vesktop still needs to provide the native node module for Wayland
