+++ src/main/mainWindow.ts
@@ -26,7 +26,6 @@ import { initArRPC } from "./arrpc";
 import { CommandLine } from "./cli";
 import { BrowserUserAgent, DEFAULT_HEIGHT, DEFAULT_WIDTH, MIN_HEIGHT, MIN_WIDTH } from "./constants";
 import { AppEvents } from "./events";
-import { initNativeIdle } from "./idle";
 import { darwinURL } from "./index";
 import { sendRendererCommand } from "./ipcCommands";
 import { Settings, State, VencordSettings } from "./settings";
@@ -37,6 +36,7 @@ import { makeLinksOpenExternally } from "./utils/makeLinksOpenExternally";
 import { applyDeckKeyboardFix, askToApplySteamLayout, isDeckGameMode } from "./utils/steamOS";
 import { downloadVencordFiles, ensureVencordFiles } from "./utils/vencordLoader";
 import { VENCORD_FILES_DIR } from "./vencordFilesDir";
+import { initWaylandIdleHandler } from "./waylandIdle";
 
 let isQuitting = false;
 
@@ -400,7 +400,6 @@ function createMainWindow() {
     initSpellCheck(win);
     initDevtoolsListeners(win);
     initStaticTitle(win);
-    initNativeIdle(win);
 
     win.webContents.setUserAgent(BrowserUserAgent);
 
@@ -442,6 +441,8 @@ export async function createWindows() {
         // SteamOS letterboxes and scales it terribly, so just full screen it
         if (isDeckGameMode) splash.setFullScreen(true);
     }
+
+    initWaylandIdleHandler();
 
     await ensureVencordFiles();
     runVencordMain();
