+++ src/preload/VesktopNative.ts
@@ -37,12 +37,7 @@ export const VesktopNative = {
         openUpdater: () => invoke<void>(IpcEvents.UPDATER_OPEN)
     },
     powerMonitor: {
-        onIdlePowerEvent: (cb: () => void) => ipcRenderer.on(IpcEvents.IDLE_POWER_EVENT, cb),
-        onNoIdlePowerEvent: (cb: () => void) => ipcRenderer.on(IpcEvents.NO_IDLE_POWER_EVENT, cb),
-        getSystemIdleTimeMs: () => invoke<number>(IpcEvents.GET_IDLE_TIME_MS),
-        isWaylandIdle: () => sendSync<boolean>(IpcEvents.IS_WAYLAND_IDLE),
-        isSuspended: () => sendSync<boolean>(IpcEvents.IS_SUSPENDED),
-        isLocked: () => sendSync<boolean>(IpcEvents.IS_LOCKED)
+        isWaylandIdle: () => sendSync<boolean>(IpcEvents.IS_WAYLAND_IDLE)
     },
     autostart: {
         isEnabled: () => sendSync<boolean>(IpcEvents.AUTOSTART_ENABLED),
