+++ src/shared/IpcEvents.ts
@@ -60,12 +60,7 @@ export const enum IpcEvents {
 
     CHOOSE_USER_ASSET = "VCD_CHOOSE_USER_ASSET",
 
-    GET_IDLE_TIME_MS = "VCD_GET_IDLE_TIME_MS",
-    IS_WAYLAND_IDLE = "VCD_IS_WAYLAND_IDLE",
-    IS_SUSPENDED = "VCD_IS_SUSPENDED",
-    IS_LOCKED = "VCD_IS_LOCKED",
-    IDLE_POWER_EVENT = "VCD_IDLE_POWER_EVENT",
-    NO_IDLE_POWER_EVENT = "VCD_NO_IDLE_POWER_EVENT"
+    IS_WAYLAND_IDLE = "VCD_IS_WAYLAND_IDLE"
 }
 
 export const enum UpdaterIpcEvents {
